fix: refactored the project structure properly
